+++ Extension/src/background/api/ui/icons.ts
@@ -86,23 +86,22 @@ class IconsApi {
      * @param frameData The information from {@link FrameData} is needed
      * to estimate the current status of the background extension
      * in the specified tab.
-     * @param frameData.documentAllowlisted Is website allowlisted.
-     * @param frameData.applicationFilteringDisabled Is app filtering disabled globally.
-     * @param frameData.totalBlockedTab Number of blocked requests.
      */
     public async updateTabAction(
         tabId: number,
-        {
+        frameData: FrameData,
+    ): Promise<void> {
+        try {
+            await this.resetPromoIconIfAny(tabId, frameData);
+        } catch { /* do nothing */ }
+
+        const {
             documentAllowlisted,
             applicationFilteringDisabled,
             totalBlockedTab,
-        }: FrameData,
-    ): Promise<void> {
-        const isDisabled = documentAllowlisted || applicationFilteringDisabled;
+        } = frameData;
 
-        try {
-            await this.setPromoIconIfAny();
-        } catch { /* do nothing */ }
+        const isDisabled = documentAllowlisted || applicationFilteringDisabled;
 
         // Determine extension's action new state based on the current tab state
         const icon = await this.pickIconVariant(isDisabled);
@@ -211,8 +210,9 @@ class IconsApi {
     }
 
     /**
-     * Fetches the current icon variants from the promo notification api, if any.
-     * Does nothing if the icon variants are already set.
+     * If promo icons variants are not set,
+     * fetches icon variants from the promo notification api (if any),
+     * otherwise does nothing.
      */
     private async setPromoIconIfAny(): Promise<void> {
         if (this.promoIcons) {
@@ -221,6 +221,23 @@ class IconsApi {
         const notification = await promoNotificationApi.getCurrentNotification();
         if (notification && notification.icons) {
             this.setPromoIcons(notification.icons);
+        }
+    }
+
+    /**
+     * Always fetches icon variants from the promo notification api,
+     * and sets the promo icons if any,
+     * otherwise promo icon is dismissed.
+     *
+     * @param tabId Tab's id.
+     * @param frameData Tab's {@link FrameData}.
+     */
+    private async resetPromoIconIfAny(tabId: number, frameData: FrameData): Promise<void> {
+        const notification = await promoNotificationApi.getCurrentNotification();
+        if (notification && notification.icons) {
+            this.setPromoIcons(notification.icons);
+        } else {
+            await this.dismissPromoIcon(tabId, frameData);
         }
     }
 }